# public/swagger.json
{
	"swagger": "2.0",
	"info": {
		"version": "1.0.0",
		"title": "TellForm API",
		"contact": {
			"name": "TellForm Team",
			"email": "team@tellform.com"
		}
	},
...
	"host": "api.tellform.com",
	"basePath": "/",
	"schemes": [
		"http",
		"https"
	],
...
	"securityDefinitions": {
		"api_key": {
			"type": "apiKey",
			"name": "apikey",
			"in": "header"
		}
	},
	"paths": {
...
		"/users/me": {
			"get": {
				"tags": [
					"user"
				],
				"summary": "Retrieve current User",
				"description": "",
				"operationId": "getUser",
				"produces": [
					"application/json"
				],
				"responses": {
					"500": {
						"description": "Could not Update User"
					},
					"401": {
						"description": "User is not Signed in"
					},
					"403": {
						"description": "User is not Authorized"
					},
					"404": {
						"description": "User does not exsit"
					},
					"200": {
						"description": "successful operation",
						"schema": {
							"$ref": "#/definitions/User"
						}
					}
				},
				"security": [{
					"api_key": [

					]
				}]
			}
		},
...
	},
	"definitions": {
		"User": {
			"type": "object",
			"required": [
				"language",
				"email",
				"username"
			],
			"properties": {
				"firstName": {
					"type": "string",
					"description": "First name of User",
					"example": "John"
				},
				"lastName": {
					"type": "string",
					"description": "First name of User",
					"example": "Doe"
				},
				"language": {
					"type": "string",
					"enum": [
						"en",
						"fr",
						"es",
						"it",
						"de"
					],
					"default": "en",
					"required": "User must have a language",
					"description": "Language of User (for internationalization)",
					"example": "fr"
				},
				"email": {
					"type": "string",
					"format": "email",
					"required": "User email cannot be blank",
					"unique": "true",
					"description": "Email of User",
					"example": "john.doe@somewhere.com"
				},
				"username": {
					"type": "string",
					"required": "Username cannot be blank",
					"unique": "true",
					"description": "Username of User",
					"example": "johndoe"
				},
				"roles": {
					"type": "array",
					"items": {
						"type": "string",
						"enum": [
							"user",
							"admin",
							"superuser"
						]
					},
					"default": [
						"user"
					],
					"description": "Security Roles of User"
				}
			},
			"lastModified": {
				"type": "date",
				"description": "Date that user was last modified",
				"example": "2016-08-26T20:19:30.146Z"
			},
			"created": {
				"type": "date",
				"description": "Date that user was created on",
				"example": "5dHuKJgeCZmFOdJTnmg0lrxApmz0tbbBrM59rTv4k79"
			},
			"resetPasswordToken": {
				"type": "string",
				"description": "Reset password token  of User",
				"example": "5dHuKJgeCZmFOdJTnmg0lrxApmz0tbbBrM59rTv4k79"
			},
			"resetPasswordExpires": {
				"type": "date",
				"example": "2016-08-26T20:19:30.146Z",
				"description": "Date that the User's password reset token expires"
			},
			"token": {
				"type": "string",
				"description": "Verification token of User",
				"example": "5dHuKJgeCZmFOdJTnmg0lrxApmz0tbbBrM59rTv4k79"
			},
			"apiKey": {
				"type": "string",
				"unique": true,
				"index": true,
				"sparse": true,
				"description": "API Key of User",
				"example": "5dHuKJgeCZmFOdJTnmg0lrxApmz0tbbBrM59rTv4k79"
			}
		},
...
	}
}